Fix voice session architecture boundary
Keep shared session state types in the React layer so state providers do not depend on the UI components that render them.
